tlscontext: fix the type of ssl_error
ERR_get_error() returns an unsigned long (gulong).

+++ lib/tlscontext.c
@@ -310,7 +310,7 @@ tls_context_setup_session(TLSContext *self)
 {
   SSL *ssl;
   TLSSession *session;
-  gint ssl_error;
+  gulong ssl_error;
   long ssl_options;
 
   if (!self->ssl_ctx)
